[WIP] move navigation.metrics package into the internal package (#1926)
