+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/feedback/FeedbackAdapter.java
@@ -5,17 +5,17 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import androidx.recyclerview.widget.RecyclerView;
-
 import com.mapbox.services.android.navigation.ui.v5.R;
 
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.mapbox.services.android.navigation.v5.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_CONFUSING_INSTRUCTION;
-import static com.mapbox.services.android.navigation.v5.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_NOT_ALLOWED;
-import static com.mapbox.services.android.navigation.v5.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_ROAD_CLOSED;
-import static com.mapbox.services.android.navigation.v5.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_ROUTING_ERROR;
+import androidx.recyclerview.widget.RecyclerView;
+
+import static com.mapbox.services.android.navigation.v5.internal.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_CONFUSING_INSTRUCTION;
+import static com.mapbox.services.android.navigation.v5.internal.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_NOT_ALLOWED;
+import static com.mapbox.services.android.navigation.v5.internal.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_ROAD_CLOSED;
+import static com.mapbox.services.android.navigation.v5.internal.navigation.metrics.FeedbackEvent.FEEDBACK_TYPE_ROUTING_ERROR;
 
 public class FeedbackAdapter extends RecyclerView.Adapter<FeedbackViewHolder> {
