+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/DepartEventHandler.java
@@ -3,8 +3,8 @@ package com.mapbox.services.android.navigation.v5.internal.navigation;
 import android.content.Context;
 
 import com.mapbox.services.android.navigation.v5.internal.location.MetricsLocation;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.SessionState;
-import com.mapbox.services.android.navigation.v5.routeprogress.MetricsRouteProgress;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.SessionState;
+import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress;
 
 class DepartEventHandler {
