+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationFeedbackEvent.java
@@ -1,11 +1,11 @@
-package com.mapbox.services.android.navigation.v5.navigation.metrics;
+package com.mapbox.services.android.navigation.v5.internal.navigation.metrics;
 
 import android.annotation.SuppressLint;
 import android.location.Location;
 
 import androidx.annotation.NonNull;
 
-import com.mapbox.services.android.navigation.v5.routeprogress.MetricsRouteProgress;
+import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress;
 
 @SuppressLint("ParcelCreator")
 class NavigationFeedbackEvent extends NavigationEvent {
