+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/internal/navigation/DepartEventFactoryTest.java
@@ -1,8 +1,8 @@
 package com.mapbox.services.android.navigation.v5.internal.navigation;
 
 import com.mapbox.services.android.navigation.v5.internal.location.MetricsLocation;
-import com.mapbox.services.android.navigation.v5.navigation.metrics.SessionState;
-import com.mapbox.services.android.navigation.v5.routeprogress.MetricsRouteProgress;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.SessionState;
+import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress;
 
 import org.junit.Test;
